Added support for inline comments. Added support for concat. Added parsing support for ByVal and ByRef. Added support for Set
