+++ lib/scripting/post-process-assign.ts
@@ -21,8 +21,14 @@ import { Expression, ExpressionStatement, MemberExpression } from 'estree';
 import { Token } from 'moo';
 import * as estree from './estree';
 
-export function stmt(result: [MemberExpression, null, Token, null, Expression]): ExpressionStatement {
+export function stmt1(result: [MemberExpression, null, Token, null, Expression]): ExpressionStatement {
 	const left = result[0];
 	const right = result[4];
+	return estree.assignmentExpressionStatement(left, '=', right);
+}
+
+export function stmt2(result: [Token, null, MemberExpression, null, Token, null, Expression]): ExpressionStatement {
+	const left = result[2];
+	const right = result[6];
 	return estree.assignmentExpressionStatement(left, '=', right);
 }
