+++ lib/scripting/post-process-sub.ts
@@ -21,20 +21,11 @@ import { BlockStatement, FunctionDeclaration, Identifier } from 'estree';
 import { Token } from 'moo';
 import * as estree from './estree';
 
-export function stmt1(
-	result: [Token, null, Token, null, Identifier, null, Identifier[], null, BlockStatement, Token, null, Token, null],
+export function stmt(
+	result: [Token[], Token, null, Identifier, null, Identifier[], null, BlockStatement, Token, null, Token, null],
 ): FunctionDeclaration {
-	const name = result[4];
-	const params = result[6] || [];
-	const body = result[8];
-	return estree.functionDeclaration(name, params, body);
-}
-
-export function stmt2(
-	result: [Token, null, Identifier, null, Identifier[], null, BlockStatement, Token, null, Token, null],
-): FunctionDeclaration {
-	const name = result[2];
-	const params = result[4] || [];
-	const body = result[6];
+	const name = result[3];
+	const params = result[5] || [];
+	const body = result[7];
 	return estree.functionDeclaration(name, params, body);
 }
